requirements.txt
Flask
Flask-HTTPAuth
python-dotenv
